+++ app/scripts/HiGlassComponent.js
@@ -5010,6 +5010,7 @@ class HiGlassComponent extends React.Component {
         // `useCSSTransforms` (it's default `true`)
         // and set `measureBeforeMount={true}`.
         useCSSTransforms={this.mounted}
+        verticalCompact={this.state.viewConfig.compactLayout}
         width={this.state.width}
       >
         {this.tiledAreas}
@@ -5103,6 +5104,7 @@ class HiGlassComponent extends React.Component {
 HiGlassComponent.defaultProps = {
   options: {},
   zoomFixed: false,
+  compactLayout: true,
 };
 
 HiGlassComponent.propTypes = {
@@ -5111,6 +5113,7 @@ HiGlassComponent.propTypes = {
   viewConfig: PropTypes.oneOfType([PropTypes.string, PropTypes.object])
     .isRequired,
   zoomFixed: PropTypes.bool,
+  compactLayout: PropTypes.bool,
 };
 
 export default HiGlassComponent;