+++ app/scripts/HiGlassComponent.js
@@ -3071,11 +3071,12 @@ class HiGlassComponent extends React.Component {
    * @param track: A view with tracks.
    */
   addCallbacks(viewUid, track) {
+    const trackInfo = this.getTrackInfo(track.type);
     if (
       track.type === 'viewport-projection-center' ||
       track.type === 'viewport-projection-horizontal' ||
       track.type === 'viewport-projection-vertical' ||
-      (track.options && track.options.addCallbacks)
+      (trackInfo && trackInfo.projection)
     ) {
       const fromView = track.fromViewUid;