+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - Allow implementing plugin viewport projection tracks using `config.projection`.
+- Change the default value of `zoomLimits` to `[1, null]`, which means highest zoom resolution is 1bp.
 
 _[Detailed changes since v1.11.5](https://github.com/higlass/higlass/compare/v1.11.5...develop)_